Updated dev version to 0.8
- Added the ability to send through various params to Add_Dir
- Added new routing (thanks to midraal for PR)
- Added Remove_From_Table function
- Renamed Run_As_Kodi to Get_ID and changed function behaviour slightly
- Split up add-on based functions into new addon module
- Added some more documentation to the README
