+++ script.module.python.koding.aio/Changelog.txt
@@ -1,21 +1,37 @@
-v.0.6.5:
+v.0.8:
+- Added the ability to send through various params to Add_Dir
+- Added new routing (thanks to midraal for PR)
+- Added Remove_From_Table function
+- Renamed Run_As_Kodi to Get_ID and changed function behaviour slightly
+- Split up add-on based functions into new addon module
+- Added some more documentation to the README
+
+v.0.7.5:
+- Rezipped, had used 7zip think that caused problems.
+
+v.0.7.4:
+- Fixed download dialog not auto-closing once complete. Last push didn't work correctly
+
+v.0.7.3:
+- Fixed the download function, wasn't automatically closing dialog once complete.
+
+v.0.7.2:
+- Tidied up the getSetting functions.
+- Fixed Create_Paths function.
+- Fixed the tutorials on Windows platforms, required url sending through as urlencoded then unencoded.
+
+v.0.7.1:
+- Added Run_As_Kodi() function which hopefully should fix Android Fresh_Install()
+- Improved Open_Settings(), can now focus on a specific item and also send an automated click on the focused item.
+
+v.0.7:
 - Added Addon_Info() function to return various built in add-on values.
 - Added Addon_Setting() function to set/retrieve an add-on setting.
 - Added new Add-on section to the README.
-
-v.0.6.4:
 - Added subprocess import for fresh_install.
-
-v.0.6.3:
 - Typo in Android check for fresh_install, hopefully fixed.
-
-v.0.6.2:
 - Added Addon_Install() function to install add-ons.
-
-v.0.6.1:
 - Added Addon_Genre() function to return a dict of all specific genre type add-ons found at the Add-on Portal.
-
-v.0.6.1:
 - Fixed unicode issue when sending through list of addons to Toggle_Addons()
 
 v.0.6:
