+++ script.module.python.koding.aio/addon.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <addon id="script.module.python.koding.aio"
    name="Python Koding AIO"
-   version="0.6.5"
+   version="0.8"
    provider-name="TotalRevolution, noobsandnerds">
   <requires>
     <import addon="xbmc.python" version="2.1.0"/>
@@ -17,8 +17,8 @@
     <language></language>
     <license>Creative Commons 4.0-NC-ND</license>
     <forum>http://noobsandnerds.com/support/showthread.php?tid=17614</forum>
-    <website>http://noobsandnerds.com</website>
-    <news>Added lots of new functions and cleaned up README. See Changelog.txt for full details.</news>
+    <website>http://totalrevolution.tv</website>
+    <news>New routing added</news>
     <source>https://github.com/totalrevolution/totalrevolution/tree/master/script.module.python.koding.aio</source>
   </extension>  
 </addon>
