+++ script.module.python.koding.aio/lib/koding/database.py
@@ -172,9 +172,23 @@ os.remove(addon_db_path)
     sql_string = sql_string[:-2]
     sql_string += ");"
     _execute_db_string(sql_string)
-
-
+#----------------------------------------------------------------
+# TUTORIAL #
 def Remove_Table(table):
+    """
+Use with caution, this will completely remove a database table and
+all of it's contents. The only database you can access with this command
+is your add-ons own db file called database.db
+
+CODE:  Remove_Table(table)
+
+AVAILABLE PARAMS:
+
+    (*) table  -  This is the name of the table you want to permanently delete.
+
+EXAMPLE CODE:
+koding.Remove_Table('my_test_table')
+~"""
     sql_string = "DROP TABLE IF EXISTS %s;" % table
     _execute_db_string(sql_string)
 #----------------------------------------------------------------
